fix unlikely-type-argument warning

+++ h2/src/test/org/h2/test/store/TestStreamStore.java
@@ -339,7 +339,7 @@ public class TestStreamStore extends TestBase {
         assertEquals(10, map.size());
         assertEquals(10, tests.get());
         for (int i = 0; i < 10; i++) {
-            map.containsKey(i);
+            map.containsKey((long)i);
         }
         assertEquals(20, tests.get());
         store = new StreamStore(map);
@@ -354,7 +354,7 @@ public class TestStreamStore extends TestBase {
         assertEquals(15, store.getNextKey());
         assertEquals(15, map.size());
         for (int i = 0; i < 15; i++) {
-            map.containsKey(i);
+            map.containsKey((long)i);
         }
     }
